requirements.txt
fastapi
uvicorn
httpx
pyyaml
psutil
pytest
pytest-asyncio